requirements.txt
python-pptx
markdown
pypinyin